requirements.txt
certifi==2018.11.29
chardet==3.0.4
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.1.7
django-anymail==5.0
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-recaptcha==2.0.2
django-registration==3.0
django-upload-validator==1.0.2
gunicorn==19.9.0
idna==2.8
Pillow==5.4.1
psycopg2==2.7.7
PyPDF2==1.26.0
python-magic==0.4.15
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Wand==0.5.1
